Synchronizing DyadicDatasetGraph::begin and breaking out Pair.OfSameType into its own compilation unit
